04-analytics-engineering/requirements-airflow.txt
apache-airflow==2.10.3
apache-airflow-providers-google==10.22.0
